StronglyTypedInterviewEvaluator: changed type from int to long in UpdateIntAnswer method KP-3738
--HG--
branch : csharp-business-language


Former-commit-id: ff1c36e3233fb246df0b15f11cdee4fce8f2fc14

+++ src/Core/SharedKernels/DataCollection/ExpressionProcessing/IInterviewExpressionState.cs
@@ -5,7 +5,7 @@ namespace WB.Core.SharedKernels.ExpressionProcessing
 {
     public interface IInterviewExpressionState
     {
-        void UpdateIntAnswer(Guid questionId, decimal[] rosterVector, int answer);
+        void UpdateIntAnswer(Guid questionId, decimal[] rosterVector, long answer);
         void UpdateDecimalAnswer(Guid questionId, decimal[] rosterVector, decimal answer);
         void UpdateDateAnswer(Guid questionId, decimal[] rosterVector, DateTime answer);
         void UpdateTextAnswer(Guid questionId, decimal[] rosterVector, string answer);
